# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Limited-run physical game tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReleaseStatus = typeof ReleaseStatus[keyof typeof ReleaseStatus];


export const ReleaseStatus = {
  available: 'available',
  sold_out: 'sold_out',
  coming_soon: 'coming_soon',
} as const;

export interface Release {
  id: number;
  title: string;
  publisherId: number;
  publisherName: string;
  publisherSlug?: string;
  platforms?: string[];
  status: ReleaseStatus;
  /** @nullable */
  coverImageUrl?: string | null;
  productUrl: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  editionType?: string | null;
  /** @nullable */
  preorderCloseDate?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  soldOutAt?: string | null;
  firstSeenAt?: string;
  createdAt: string;
  updatedAt: string;
}

export interface ReleaseListResponse {
  releases: Release[];
  total: number;
}

export interface ReleaseStats {
  available: number;
  soldOut: number;
  comingSoon: number;
  totalTracked: number;
  /** @nullable */
  lastUpdated: string | null;
}

export interface Publisher {
  id: number;
  name: string;
  slug: string;
  websiteUrl: string;
  /** @nullable */
  logoUrl?: string | null;
  enabled: boolean;
  releaseCount: number;
  /** @nullable */
  lastScrapedAt?: string | null;
}

export interface Platform {
  name: string;
  releaseCount: number;
}

export interface SubscriberInput {
  email: string;
}

export interface SubscribeResult {
  message: string;
}

/**
 * @nullable
 */
export type ScrapeStatusLastRunStatus = typeof ScrapeStatusLastRunStatus[keyof typeof ScrapeStatusLastRunStatus] | null;


export const ScrapeStatusLastRunStatus = {
  running: 'running',
  success: 'success',
  error: 'error',
} as const;

export interface ScrapeStatus {
  publisherId: number;
  publisherName: string;
  publisherSlug: string;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastRunStatus?: ScrapeStatusLastRunStatus;
  /** @nullable */
  releasesFound?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface ScrapeInput {
  /**
     * If omitted, scrape all enabled publishers
     * @nullable
     */
  publisherSlug?: string | null;
}

export interface ScrapeResult {
  message: string;
  publishersTriggered: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListReleasesParams = {
status?: ListReleasesStatus;
platform?: string;
publisher?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListReleasesStatus = typeof ListReleasesStatus[keyof typeof ListReleasesStatus];


export const ListReleasesStatus = {
  available: 'available',
  sold_out: 'sold_out',
  coming_soon: 'coming_soon',
} as const;

export type ListAvailableReleasesParams = {
platform?: string;
publisher?: string;
};

export type ListSoldOutReleasesParams = {
platform?: string;
publisher?: string;
limit?: number;
};

export type ListComingSoonReleasesParams = {
platform?: string;
publisher?: string;
};