# lib/api-zod/src/generated/types/release.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Limited-run physical game tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ReleaseStatus } from './releaseStatus';

export interface Release {
  id: number;
  title: string;
  publisherId: number;
  publisherName: string;
  publisherSlug?: string;
  platforms?: string[];
  status: ReleaseStatus;
  /** @nullable */
  coverImageUrl?: string | null;
  productUrl: string;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  editionType?: string | null;
  /** @nullable */
  preorderCloseDate?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  soldOutAt?: string | null;
  firstSeenAt?: string;
  createdAt: string;
  updatedAt: string;
}