# lib/api-zod/src/generated/types/scrapeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Limited-run physical game tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeInput {
  /**
     * If omitted, scrape all enabled publishers
     * @nullable
     */
  publisherSlug?: string | null;
}