# lib/api-zod/src/generated/types/scrapeStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Limited-run physical game tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapeStatusLastRunStatus } from './scrapeStatusLastRunStatus';

export interface ScrapeStatus {
  publisherId: number;
  publisherName: string;
  publisherSlug: string;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastRunStatus?: ScrapeStatusLastRunStatus;
  /** @nullable */
  releasesFound?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}