# lib/api-zod/src/generated/types/subscribeResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Limited-run physical game tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeResult {
  message: string;
}